code refactored and docker file changed
