+++ requirements.txt
@@ -41,3 +41,4 @@ traitlets==4.3.2
 wcwidth==0.1.7
 Werkzeug==0.15.4
 Flask-HTTPBasicAuth
+Flask-Fixtures
